# src/main/resources/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Fuzzer demo",
    "version": "1.0",
    "title": "Fuzzer demo"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
    "/foo": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "authToken",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "aa",
            "description": "aa",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ReloadSvcResponse"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/bar": {
      "post": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "authToken",
            "in": "header",
            "required": true,
            "type": "string"
          },
          {
            "in": "body",
            "name": "bb",
            "description": "bb",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "201": {
            "description": "CREATED",
            "schema": {
              "$ref": "#/definitions/SuccessfulAddCreditCardResponse"
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "get": {
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "x-auth-token",
            "description": "encrypted auth token",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    }
  },
  "definitions": {
  }
}